refactor: create cost center repository
jira: B2BTEAM-1382
